refactor: NoxStorage namespace
