+++ src/enums/Playlist.ts
@@ -10,3 +10,11 @@ export const SearchRegex: { [key: string]: { regex: RegExp; text: string } } = {
   albumMatch: { regex: /Album:(.+)/, text: 'Album:' },
   cachedMatch: { regex: /Cached:/, text: 'Cached:' },
 };
+
+export enum SORT_OPTIONS {
+  TITLE = 'title',
+  ARTIST = 'artist',
+  ALBUM = 'album',
+  DATE = 'date',
+  PREVIOUS_ORDER = 'previous',
+}
