+++ src/objects/Playlist.ts
@@ -30,18 +30,3 @@ export const getPlaylistUniqBVIDs = (playlist: NoxMedia.Playlist) => {
     )
   );
 };
-
-export const updatePlaylistSongs = (
-  playlist: NoxMedia.Playlist,
-  addSongs: Array<NoxMedia.Song> = [],
-  removeSongs: Array<NoxMedia.Song> = []
-) => {
-  const playlistSongsId = playlist.songList.map(v => v.id);
-  const removeSongsId = removeSongs.map(v => v.id);
-  // FI"FO".
-  playlist.songList = addSongs
-    .filter(v => !playlistSongsId.includes(v.id))
-    .concat(playlist.songList)
-    .filter(v => !removeSongsId.includes(v.id));
-  return playlist;
-};
