+++ src/services/PlaybackService.ts
@@ -8,7 +8,6 @@ import { DeviceEventEmitter, Platform } from 'react-native';
 import { NULL_TRACK } from '../objects/Song';
 import { parseSongR128gain } from '../utils/SongOperations';
 import { initBiliHeartbeat } from '../utils/Bilibili/BiliOperate';
-import type { NoxStorage } from '../types/storage';
 import { logger } from '../utils/Logger';
 import noxPlayingList, { getNextSong } from '../stores/playingList';
 import { NoxRepeatMode } from '../enums/RepeatMode';
