+++ src/types/storage.d.ts
@@ -2,85 +2,86 @@
 import { ColorSchemeName } from 'react-native';
 
 import { EXPORT_OPTIONS } from '../enums/Sync';
+declare global {
+  namespace NoxStorage {
+    export interface PlayerSettingDict {
+      playMode: string;
+      defaultPlayMode: string;
+      defaultVolume: number;
 
-declare namespace NoxStorage {
-  export interface PlayerSettingDict {
-    playMode: string;
-    defaultPlayMode: string;
-    defaultVolume: number;
+      autoRSSUpdate: boolean;
+      skin: string;
+      parseSongName: boolean;
+      keepSearchedSongListWhenPlaying: boolean;
+      settingExportLocation: EXPORT_OPTIONS;
+      personalCloudIP: string;
+      personalCloudID: string;
+      noxVersion: string;
+      noxCheckedVersion: string;
 
-    autoRSSUpdate: boolean;
-    skin: string;
-    parseSongName: boolean;
-    keepSearchedSongListWhenPlaying: boolean;
-    settingExportLocation: EXPORT_OPTIONS;
-    personalCloudIP: string;
-    personalCloudID: string;
-    noxVersion: string;
-    noxCheckedVersion: string;
+      hideCoverInMobile: boolean;
+      loadPlaylistAsArtist: boolean;
+      sendBiliHeartbeat: boolean;
+      // TODO: implement this feature
+      noCookieBiliSearch: boolean;
+      // TODO: implement this feature
+      dataSaver: boolean;
+      fastBiliSearch: boolean;
+      noInterruption: boolean;
+      updateLoadedTrack: boolean;
+      r128gain: boolean;
+      prefetchTrack: boolean;
+      // TODO: implement this feature
+      chatGPTResolveSongName: boolean;
+      trackCoverArtCard: boolean;
+      suggestedSkipLongVideo: boolean;
+      wavyProgressBar: boolean;
 
-    hideCoverInMobile: boolean;
-    loadPlaylistAsArtist: boolean;
-    sendBiliHeartbeat: boolean;
-    // TODO: implement this feature
-    noCookieBiliSearch: boolean;
-    // TODO: implement this feature
-    dataSaver: boolean;
-    fastBiliSearch: boolean;
-    noInterruption: boolean;
-    updateLoadedTrack: boolean;
-    r128gain: boolean;
-    prefetchTrack: boolean;
-    // TODO: implement this feature
-    chatGPTResolveSongName: boolean;
-    trackCoverArtCard: boolean;
-    suggestedSkipLongVideo: boolean;
-    wavyProgressBar: boolean;
+      appID: string;
+      language?: string;
+      cacheSize: number;
+      [key: string]: any;
+    }
 
-    appID: string;
-    language?: string;
-    cacheSize: number;
-    [key: string]: any;
-  }
+    export interface PlayerStorageObject {
+      settings: PlayerSettingDict;
+      playlistIds: Array<string>;
+      playlists: { [key: string]: NoxMedia.Playlist };
+      lastPlaylistId: [string, string];
+      searchPlaylist: NoxMedia.Playlist;
+      favoriPlaylist: NoxMedia.Playlist;
+      playerRepeat: string;
+      skin: NoxTheme.Style;
+      skins: any[];
+      // site: set-cookie header
+      cookies: { [key: string]: string };
+      lyricMapping: Map<string, NoxMedia.LyricDetail>;
+      language?: string;
+      lastPlayDuration: number;
+      colorScheme: ColorSchemeName;
+    }
 
-  export interface PlayerStorageObject {
-    settings: PlayerSettingDict;
-    playlistIds: Array<string>;
-    playlists: { [key: string]: NoxMedia.Playlist };
-    lastPlaylistId: [string, string];
-    searchPlaylist: NoxMedia.Playlist;
-    favoriPlaylist: NoxMedia.Playlist;
-    playerRepeat: string;
-    skin: NoxTheme.Style;
-    skins: any[];
-    // site: set-cookie header
-    cookies: { [key: string]: string };
-    lyricMapping: Map<string, NoxMedia.LyricDetail>;
-    language?: string;
-    lastPlayDuration: number;
-    colorScheme: ColorSchemeName;
-  }
+    export interface initializedResults {
+      currentPlayingList: NoxMedia.Playlist;
+      currentPlayingID: string;
+      playlists: { [key: string]: NoxMedia.Playlist };
+      storedPlayerSetting: NoxStorage.PlayerSettingDict;
+      cookies: { [key: string]: string };
+      language?: string;
+      lastPlayDuration: number;
+      playbackMode: string;
+    }
 
-  export interface initializedResults {
-    currentPlayingList: NoxMedia.Playlist;
-    currentPlayingID: string;
-    playlists: { [key: string]: NoxMedia.Playlist };
-    storedPlayerSetting: NoxStorage.PlayerSettingDict;
-    cookies: { [key: string]: string };
-    language?: string;
-    lastPlayDuration: number;
-    playbackMode: string;
-  }
+    export interface R128Dict {
+      [key: string]: number | null;
+    }
 
-  export interface R128Dict {
-    [key: string]: number | null;
-  }
+    export interface ABDict {
+      [key: string]: [number, number];
+    }
 
-  export interface ABDict {
-    [key: string]: [number, number];
-  }
-
-  export interface DownloadDict {
-    [key: string]: Promise<void | string>;
+    export interface DownloadDict {
+      [key: string]: Promise<void | string>;
+    }
   }
 }
